rename redux object to a more logical name
