+++ js/src/redux/actions/index.js
@@ -13,7 +13,7 @@ export * from "./estimatedReadingTime";
 export * from "./focusKeyword";
 export * from "./markerButtons";
 export * from "./markerPauseStatus";
-export * from "./persistentDismissableAlert";
+export * from "./dismissedAlerts";
 export * from "./primaryTaxonomies";
 export * from "./schemaTab";
 export * from "./schemaBlocks";
