+++ js/src/redux/selectors/index.js
@@ -8,7 +8,7 @@ export * from "./facebookEditor";
 export * from "./fallbacks";
 export * from "./focusKeyPhrase";
 export * from "./marker";
-export * from "./persistentDismissableAlert";
+export * from "./dismissedAlerts";
 export * from "./preferences";
 export * from "./primaryTaxonomies";
 export * from "./results";
